+++ server/src/partition_server.cpp
@@ -1,6 +1,6 @@
 #include "../include/partition_server.h"
 
-Partition_Server::Partition_Server(uint32_t port) : Server(port) {
+Partition_Server::Partition_Server(uint16_t port) : Server(port) {
 
 }
